chore: rename api.queue to api.queue_context to help sphinx
